Description of the poem
The poem is about an old computer. (First line) The computer is being turned on. (Second line) The computer's screen is being reflected on the person's eye. (Third line) Since the reading talks about how amazed and entertained people were with the early invention of the computer because of its four properties (procedural, participatory, spatial, and encyclopedic), the computer's "eye" and a person's eye interlock because both are facing each other and the person is completely immersed by what is being displayed on the screen. Furthermore, the computer is displaying information to the person, and the person is responding back, which is what is referred as this "conversation" happening. (Fourth line) The computer knows and has seen much more than a human. (Fifth line) Old computers had constant problems and took forever to fix it (as mentioned in the reading), so the lifespan of a computer is not as long as a healthy human's eye.